Use github API to find versions instead of Downloads page.
This uses Github API to fetch all tags and use the tagging pattern of
vMAJOR.MINOR.PATCH[PRE_RELEASE] style to find all the version tags and then
filter them based on the series we are tracking.

This has an added advantage that we can get even pre-releases for series' that
isn't yet release like 3.10 right now.